Move utility mehods and fix types
