+++ core/src/main/scala-2/chisel3/PrintfMacros.scala
@@ -17,7 +17,7 @@ object PrintfMacrosCompat {
   ): c.Tree = {
     import c.universe._
     _checkFormatString(c)(fmt)
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("chisel3.printf.printfWithReset"))
+    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("printfWithReset"))
     q"$apply_impl_do(_root_.chisel3.Printable.pack($fmt, ..$data))($sourceInfo)"
   }
 
@@ -52,4 +52,38 @@ object PrintfMacrosCompat {
       case _ =>
     }
   }
+
+  private[chisel3] def printfWithReset(
+    pable: Printable
+  )(
+    implicit sourceInfo: SourceInfo
+  ): chisel3.printf.Printf = {
+    var printfId: chisel3.printf.Printf = null
+    when(!Module.reset.asBool) {
+      printfId = printfWithoutReset(pable)
+    }
+    printfId
+  }
+
+  private[chisel3] def printfWithoutReset(
+    pable: Printable
+  )(
+    implicit sourceInfo: SourceInfo
+  ): chisel3.printf.Printf = {
+    val clock = Builder.forcedClock
+    val printfId = new chisel3.printf.Printf(pable)
+
+    Printable.checkScope(pable)
+
+    pushCommand(chisel3.internal.firrtl.ir.Printf(printfId, sourceInfo, clock.ref, pable))
+    printfId
+  }
+
+  private[chisel3] def printfWithoutReset(
+    fmt:  String,
+    data: Bits*
+  )(
+    implicit sourceInfo: SourceInfo
+  ): chisel3.printf.Printf =
+    printfWithoutReset(Printable.pack(fmt, data: _*))
 }
