+++ core/src/main/scala-2/chisel3/VerificationStatementMacros.scala
@@ -4,13 +4,38 @@ package chisel3
 
 import chisel3.experimental.SourceInfo
 import chisel3.PrintfMacrosCompat._
+import chisel3.internal.firrtl.ir._
+import chisel3.layer.block
+import chisel3.layers
+import chisel3.util.circt.IfElseFatalIntrinsic
+import chisel3.internal.Builder.pushCommand
+import chisel3.internal._
+
 import scala.language.experimental.macros
 import scala.reflect.macros.blackbox
 import scala.reflect.macros.blackbox.Context
 import scala.annotation.nowarn
 
 object VerifStmtMacrosCompat {
+
   type SourceLineInfo = (String, Int)
+
+  def formatFailureMessage(
+    kind:     String,
+    lineInfo: SourceLineInfo,
+    cond:     Bool,
+    message:  Option[Printable]
+  )(
+    implicit sourceInfo: SourceInfo
+  ): Printable = {
+    val (filename, line) = lineInfo
+    val lineMsg = s"$filename:$line".replaceAll("%", "%%")
+    message match {
+      case Some(msg) =>
+        p"$kind failed: $msg\n"
+      case None => p"$kind failed at $lineMsg\n"
+    }
+  }
 
   def getLine(c: blackbox.Context): SourceLineInfo = {
     val p = c.enclosingPosition
@@ -68,6 +93,34 @@ object VerifStmtMacrosCompat {
       val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
       q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.None)($sourceInfo)"
     }
+
+    /** @group VerifPrintMacros */
+    def _applyWithSourceLinePrintable(
+      cond:    Bool,
+      line:    SourceLineInfo,
+      message: Option[Printable]
+    )(
+      implicit sourceInfo: SourceInfo
+    ): chisel3.assert.Assert = {
+      message.foreach(Printable.checkScope(_))
+      val pable = formatFailureMessage("Assertion", line, cond, message)
+      emitIfElseFatalIntrinsic(Module.clock, cond, !Module.reset.asBool, pable)
+    }
+
+    private def emitIfElseFatalIntrinsic(
+      clock:     Clock,
+      predicate: Bool,
+      enable:    Bool,
+      format:    Printable
+    )(
+      implicit sourceInfo: SourceInfo
+    ): chisel3.assert.Assert = {
+      block(layers.Verification.Assert, skipIfAlreadyInBlock = true, skipIfLayersEnabled = true) {
+        val id = Builder.forcedUserModule // It should be safe since we push commands anyway.
+        IfElseFatalIntrinsic(id, format, "chisel3_builtin", clock, predicate, enable, format.unpackArgs: _*)(sourceInfo)
+      }
+      new chisel3.assert.Assert()
+    }
   }
 
   object assume {
@@ -121,6 +174,25 @@ object VerifStmtMacrosCompat {
       val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
       q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.None)($sourceInfo)"
     }
+
+    /** @group VerifPrintMacros */
+    def _applyWithSourceLinePrintable(
+      cond:    Bool,
+      line:    SourceLineInfo,
+      message: Option[Printable]
+    )(
+      implicit sourceInfo: SourceInfo
+    ): chisel3.assume.Assume = {
+      val id = new chisel3.assume.Assume()
+      block(layers.Verification.Assume, skipIfAlreadyInBlock = true, skipIfLayersEnabled = true) {
+        message.foreach(Printable.checkScope(_))
+        when(!Module.reset.asBool) {
+          val formattedMsg = formatFailureMessage("Assumption", line, cond, message)
+          Builder.pushCommand(Verification(id, Formal.Assume, sourceInfo, Module.clock.ref, cond.ref, formattedMsg))
+        }
+      }
+      id
+    }
   }
 
   object cover {
@@ -146,6 +218,23 @@ object VerifStmtMacrosCompat {
       import c.universe._
       val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLine"))
       q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some($message))($sourceInfo)"
+    }
+
+    /** @group VerifPrintMacros */
+    def _applyWithSourceLine(
+      cond:    Bool,
+      line:    SourceLineInfo,
+      message: Option[String]
+    )(
+      implicit sourceInfo: SourceInfo
+    ): chisel3.cover.Cover = {
+      val id = new chisel3.cover.Cover()
+      block(layers.Verification.Cover, skipIfAlreadyInBlock = true, skipIfLayersEnabled = true) {
+        when(!Module.reset.asBool) {
+          Builder.pushCommand(Verification(id, Formal.Cover, sourceInfo, Module.clock.ref, cond.ref, ""))
+        }
+      }
+      id
     }
   }
 }
